refactor: extract static classes from Modifiers
