+++ src/net/sourceforge/kolmafia/PastaThrallData.java
@@ -5,6 +5,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import net.sourceforge.kolmafia.modifiers.Lookup;
 import net.sourceforge.kolmafia.modifiers.StringModifier;
 import net.sourceforge.kolmafia.objectpool.SkillPool;
 import net.sourceforge.kolmafia.preferences.Preferences;
@@ -329,7 +330,7 @@ public class PastaThrallData implements Comparable<PastaThrallData> {
   private int level;
   private String name;
   private String mods;
-  private Modifiers.Lookup modsLookup;
+  private Lookup modsLookup;
 
   public PastaThrallData(final PastaThrallType data) {
     this.data = data;
